refactoring/#17: Banner Type 추상화
