+++ src/widgets/BannerSwiper/model/bannerData.ts
@@ -1,13 +1,4 @@
-export interface Banner {
-  id: string;
-  title: string;
-  imageUrl: string;
-  href: string;
-  status?: 'ongoing' | 'ended';
-  startDate: string;
-  endDate: string;
-  buttonLabel?: string;
-}
+import { Banner } from './types';
 
 export const bannerList: Banner[] = [
   {
